todolist/requirements.txt
asgiref==3.9.2
Django==5.2.6
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
